backend/requirements.txt
fastapi
uvicorn
python-multipart
PyMuPDF
pytesseract
Pillow
python-docx